# internal/web/openapi.yaml
openapi: 3.0.3
info:
  title: goca API
  version: "1.0"
  description: |
    The REST API of **goca**, a self-hosted certificate authority. Every
    capability here also exists in the CLI and the web portal — all three
    call the same service layer, so nothing in this spec is API-only.

    ## Authentication

    Every endpoint under `/api/v1` (except `/api/v1/health`) accepts either:

    - **A bearer token** — `Authorization: Bearer goca_xxxxxxxx_...`, created
      with `goca token create` or `POST /api/v1/tokens`. This is the normal
      way to call the API from scripts and automation.
    - **A portal session cookie** — if you're already signed in to the portal
      in this browser, "Try it out" below works using that session. Session-
      authenticated `POST`/`PATCH`/`DELETE` calls also require a CSRF token,
      which this docs page is not set up to supply — use a bearer token for
      anything beyond `GET` requests here.

    Endpoints marked 🔒 **admin** require the caller's role (bearer token or
    session) to be `admin`; a `user`-role token gets a 403 from those.

    ## Downloads

    The `.../download/{file}` endpoints return raw file bytes (PEM, DER,
    a PKCS#12 or ZIP archive), not JSON — their `Content-Type` varies by the
    `file` requested. They're documented here for completeness but a browser
    "Try it out" will just show the response inline.

    ## ACME itself

    The ACME (RFC 8555) protocol endpoints under `/acme/*` are a separate,
...
servers:
...
security:
  - bearerAuth: []
  - sessionAuth: []
...
paths:
...
  /secrets:
    get:
      tags: [Secrets]
      summary: List secrets
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [kv, file, certificate] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  secrets: { type: array, items: { $ref: "#/components/schemas/Secret" } }
                  count: { type: integer }
    post:
      tags: [Secrets]
      summary: Create a secret
      description: Metadata only. A kv/file secret has no version until you POST one; a certificate secret is complete immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSecretInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Secret" }
        "400": { description: Invalid input }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "An API token from `goca token create` / `POST /tokens`, e.g. `goca_ab12cd34_...`."
    sessionAuth:
      type: apiKey
      in: cookie
      name: goca_session
      description: The portal's own session cookie. State-changing calls also need a CSRF token this docs page cannot supply — prefer a bearer token here.
...
  schemas:
...
    Secret:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, description: "path-like and unique, e.g. team-a/db/password" }
        type: { type: string, enum: [kv, file, certificate] }
        description: { type: string }
        cert_id: { type: integer, nullable: true, description: "set only for certificate secrets" }
        current_version: { type: integer, description: "0 for a certificate secret, which has no versions of its own" }
        rotation_days: { type: integer, description: "advisory; flags the secret 'rotation due' in lists" }
        disabled: { type: boolean }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        # Computed server-side: the stored labels are a JSON blob that is not
        # itself serialized, so without this an API client sees no labels.
        labels: { type: object, additionalProperties: { type: string } }
        rotation_due: { type: boolean }
    CreateSecretInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        type: { type: string, enum: [kv, file, certificate], default: kv }
        description: { type: string }
        labels: { type: object, additionalProperties: { type: string } }
        rotation_days: { type: integer }
        cert_id: { type: integer, description: "required, and only valid, when type is certificate" }
...